Release 0.3.2-alpha

+++ nf/subworkflows/ncbi/rnaseq_short/fetch_sra_fasta/main.nf
@@ -48,8 +48,10 @@ process run_fetch_sra_fasta {
     curl -o ${sra} \$output_${sra}
     fasterq-dump --skip-technical --threads 6 --split-files --seq-defline ">\\\$ac.\\\$si.\\\$ri" --fasta -O .  ./${sra}
     rm -f ${sra}
-    mkdir output
-    mv ${sra}_1.fasta output/${sra}.1
+    mkdir -p output
+    if [ -f ${sra}_1.fasta ]; then
+        mv ${sra}_1.fasta output/${sra}.1
+    fi
     if [ -f ${sra}_2.fasta ]; then
         mv ${sra}_2.fasta output/${sra}.2
     fi
